Add files via upload
The Shadownet/Carthage range did NOT have a PyYAML5.4.1 for pip install into this webapp's environment.  So we must download the tar.gz and install manaully onto the VM during build.  For a host connected to grey internet, this file will not be needed, as pip will be able to get the 5.4.1 version from internet